clippy

+++ crates/testing/tests/tests_3/byzantine_tests.rs
@@ -179,7 +179,7 @@ cross_tests!(
                 view_increment: nodes_count,
                 modifier: Arc::new(move |_pk, message_kind, transmit_type: &mut TransmitType<TestTypes>, membership: &<TestTypes as NodeType>::Membership| {
                     if let MessageKind::Consensus(SequencingMessage::General(GeneralConsensusMessage::Vote(vote))) = message_kind {
-                        *transmit_type = TransmitType::Direct(membership.leader(vote.view_number() + 1 - nodes_count as u64, None).unwrap());
+                        *transmit_type = TransmitType::Direct(membership.leader(vote.view_number() + 1 - nodes_count, None).unwrap());
                     } else {
                         {}
                     }
